feat: remove deprecated, move experimental files it's a feat version because "breaking" below v1
