+++ src/js/experimental/media-settings-popup.js
@@ -1,6 +1,6 @@
 // Work in progress
 
-import { window, document } from './utils/server-safe-globals.js';
+import { window, document } from '../utils/server-safe-globals.js';
 
 const template = document.createElement('template');
