+++ src/js/index.js
@@ -20,13 +20,11 @@ import MediaPipButton from './media-pip-button.js';
 import MediaPlayButton from './media-play-button.js';
 import MediaPlaybackRateButton from './media-playback-rate-button.js';
 import MediaPosterImage from './media-poster-image.js';
-import MediaProgressRange from './media-progress-range.js';
 import MediaSeekBackwardButton from './media-seek-backward-button.js';
 import MediaPreviewTimeDisplay from './media-preview-time-display.js';
 import MediaPreviewThumbnail from './media-preview-thumbnail.js';
 import MediaTimeRange from './media-time-range.js';
 import MediaLoadingIndicator from './media-loading-indicator.js';
-import MediaTitleElement from './media-title-element.js';
 import MediaVolumeRange from './media-volume-range.js';
 import { window } from './utils/server-safe-globals.js';
 
@@ -73,12 +71,10 @@ export {
   MediaPlayButton,
   MediaPlaybackRateButton,
   MediaPosterImage,
-  MediaProgressRange,
   MediaSeekBackwardButton,
   MediaPreviewTimeDisplay,
   MediaPreviewThumbnail,
   MediaTimeRange,
-  MediaTitleElement,
   MediaLoadingIndicator,
   MediaVolumeRange,
 };
